Rename Masthead to Hero
