+++ src/components/Hero.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import GatsbyImage from './GatsbyImage';
-import Type1 from '../components/Type1';
+import Type1 from './Type1';
 import * as CustomPropTypes from '../propTypes';
 import { COLORS } from '../styles/vars';
 import { pxToRem } from '../styles/utils';
@@ -16,8 +16,8 @@ const BackgroundImage = GatsbyImage.extend`
   opacity: ${props => props.opacity};
 `;
 
-export class Masthead extends Component {
-  static displayName = 'Masthead';
+export class Hero extends Component {
+  static displayName = 'Hero';
 
   static propTypes = {
     tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
@@ -86,7 +86,7 @@ export class Masthead extends Component {
   }
 }
 
-export default styled(Masthead)`
+export default styled(Hero)`
   background-color: ${props => props.bgColor};
   margin-bottom: 1.45rem;
   overflow: hidden;
